requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
passlib==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
starlette==0.26.1
pydantic==1.10.7
email-validator==2.0.0
python-jose==3.3.0
bcrypt==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyotp==2.8.0
itsdangerous==2.1.2
httpx==0.24.1
requests==2.28.1